Added Button with Functionality to Generate the World
Currently generates new noise maps, which you can see below the viewport if you zoom out.

This functionality can be translated to the generation of the 3d voxel world.

We can either have just the water, an initial generated world, or an empty scene before clicking the generate world button.